Statistics

+++ src/components/AdministrationFlashCards.jsx
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
+import { confirmAlert } from 'react-confirm-alert'; // Import
+import 'react-confirm-alert/src/react-confirm-alert.css'; // Import css
 import WebSpeechApi from '../utils/WebSpeechApi';
 
 import '../assets/css/administration.css';
@@ -28,7 +30,21 @@ class AdministrationFlashCards extends Component {
 
   handleDeleteCardChange(event, id) {
     event.preventDefault();
-    this.props.deleteCard(id);
+    confirmAlert({
+      title: 'Confirm to delete this flashcard?',
+      message: 'Are you sure to delete this.',
+      buttons: [
+        {
+          label: 'Yes',
+          onClick: () => {
+            this.props.deleteCard(id);
+          },
+        },
+        {
+          label: 'No',
+        },
+      ],
+    });
   }
 
   render() {
@@ -51,14 +67,14 @@ class AdministrationFlashCards extends Component {
     }
     return (
       <div className="container administration">
-        <h2>manage your flashcards</h2>
+        <h2>Manage your flashcards</h2>
         <p>
           <button
             onClick={() => this.changePage('createCard')}
             type="button"
             className="btn btn-primary"
           >
-            create new flashcard
+            Create new flashcard
           </button>
         </p>
         <table className="table table-striped">
@@ -81,7 +97,7 @@ class AdministrationFlashCards extends Component {
                     <div className="btn-group btn-group-justified">
                       <button
                           type="button"
-                          className="btn btn-default"
+                          className="btn btn-success"
                           onClick={event => this.handleUpdateCardChange(event, card.id)}
                         >
                         update
